+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v2.0.0
+
+- Bug Fix
+- Added Filewatcher. Automatically restarts the server for file changes.
+- updated `@r35007/mock-server@2.0.7`
 ## v1.0.0
 
 - Initial release